requirements.txt
flask
peewee #数据库连接软件包 sqlite connection package
python-dotenv #获取环境变量
pycryptodome #加密软件包
pyjwt #JWT软件包